+++ packages/components/badge/examples/EntityStatusBadgeChangedExample.tsx
@@ -1,0 +1,6 @@
+import React from 'react';
+import { EntityStatusBadge } from '@contentful/f36-components';
+
+export default function EntityStatusBadgeChangedExample() {
+  return <EntityStatusBadge entityStatus="changed" />;
+}