+++ packages/components/badge/examples/EntityStatusBadgeDraftExample.tsx
@@ -1,0 +1,6 @@
+import React from 'react';
+import { EntityStatusBadge } from '@contentful/f36-components';
+
+export default function EntityStatusBadgeDraftExample() {
+  return <EntityStatusBadge entityStatus="draft" />;
+}