+++ packages/components/badge/examples/EntityStatusBadgeScheduledExample.tsx
@@ -1,0 +1,6 @@
+import React from 'react';
+import { EntityStatusBadge } from '@contentful/f36-components';
+
+export default function EntityStatusBadgeScheduledExample() {
+  return <EntityStatusBadge entityStatus="changed" isScheduled />;
+}