+++ packages/website/components/LiveEditor/ComponentSource.tsx
@@ -11,7 +11,10 @@ import { LiveProvider, LiveEditor, LiveError, LivePreview } from 'react-live';
 import tokens from '@contentful/f36-tokens';
 import * as f36Components from '@contentful/f36-components';
 import { Multiselect } from '@contentful/f36-multiselect';
+import { NavList } from '@contentful/f36-navlist';
 import * as f36utils from '@contentful/f36-utils';
+import { Avatar, Variant, Size } from '@contentful/f36-avatar';
+import { Image } from '@contentful/f36-image';
 import { useForm, useController } from 'react-hook-form';
 import { MdAccessAlarm } from 'react-icons/md';
 import { Card, Button, CopyButton, Flex } from '@contentful/f36-components';
@@ -21,19 +24,27 @@ import { theme } from './theme';
 import { formatSourceCode } from './utils';
 import * as coder from '../../utils/coder';
 import FocusLock from 'react-focus-lock';
+import { DndContext } from '@dnd-kit/core';
 import {
-  SortableContainer,
-  SortableElement,
-  SortableHandle,
-} from 'react-sortable-hoc';
-import { arrayMoveImmutable } from 'array-move';
+  arrayMove,
+  horizontalListSortingStrategy,
+  SortableContext,
+  useSortable,
+  verticalListSortingStrategy,
+} from '@dnd-kit/sortable';
+import { CSS } from '@dnd-kit/utilities';
 import { format, parse, isValid } from 'date-fns';
 
 const liveProviderScope = {
   ...f36icons,
   ...f36Components,
   ...f36utils,
+  Avatar, // Remove when avatar is added to f36-components
+  Variant, // Remove when avatar is added to f36-components
+  Size, // Remove when avatar is added to f36-components
+  Image, // Remove when added to f36-components
   Multiselect, // Remove when added to f36-components
+  NavList, // Remove when added to f36-components
   css,
   f36icons,
   tokens,
@@ -44,15 +55,21 @@ const liveProviderScope = {
   useRef,
   useMemo,
   useContext,
-  // other
+  // react-hook-form
   useForm,
   useController,
+  // other
   MdAccessAlarm,
   FocusLock,
-  SortableContainer,
-  SortableElement,
-  SortableHandle,
-  arrayMoveImmutable,
+  // dnd-kit
+  arrayMove,
+  CSS,
+  DndContext,
+  horizontalListSortingStrategy,
+  SortableContext,
+  useSortable,
+  verticalListSortingStrategy,
+  // date-fns
   format,
   parse,
   isValid,
@@ -65,7 +82,7 @@ const styles = {
   error: css`
     font-family: ${tokens.fontStackMonospace};
     font-size: ${tokens.fontSizeS};
-    background: ${tokens.colorNegative};
+    background: ${tokens.red600};
     color: ${tokens.colorWhite};
     padding: ${tokens.spacingXs};
   `,