+++ packages/website/components/Playground/SandpackRenderer.tsx
@@ -89,11 +89,16 @@ export function SandpackRenderer({
       css={sandpackStyles.wrapper}
       customSetup={{
         dependencies: {
+          '@dnd-kit/core': '^6.0.0',
+          '@dnd-kit/sortable': '^7.0.0',
           react: '^17.0.0',
           'react-dom': '^17.0.0',
           'react-scripts': '^4.0.0',
           '@contentful/f36-components': '^4.0.0',
+          '@contentful/f36-avatar': '^4.0.0-alpha.0', // Remove when added to f36-components
+          '@contentful/f36-image': '^4.0.0-alpha.0', // Remove when added to f36-components
           '@contentful/f36-multiselect': '^4.0.0', // Remove when added to f36-components
+          '@contentful/f36-navlist': '^4.1.0-alpha.0', // Remove when added to f36-components
           '@contentful/f36-tokens': '^4.0.0',
           '@contentful/f36-icons': '^4.0.0',
           emotion: '^10.0.17',
@@ -101,8 +106,6 @@ export function SandpackRenderer({
           'react-hook-form': '7.22.5',
           'react-icons': '4.3.1',
           'react-focus-lock': '^2.5.2',
-          'react-sortable-hoc': '^2.0.0',
-          'array-move': '^4.0.0',
         },
       }}
       files={{